% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_utilities.R
\name{evaluate_grid_point}
\alias{evaluate_grid_point}
\title{Evaluate a parameter combination}
\usage{
evaluate_grid_point(n_edges, params, time_window, cascades, min_time = NULL,
  decay_parameter)
}
\arguments{
\item{cascades}{an object of class cascade containing node and cascade
information. See \code{\link{as_cascade_long}} and
\code{\link{as_cascade_wide}} for details.}
}
\value{
Returns an object of class \code{data.frame}
}
\description{
This the short description
}
\details{
And some details
}
\examples{

data(cascades)

}
