% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_utilities.R
\name{grid_search_eha}
\alias{grid_search_eha}
\title{Evaluate a parameter grid}
\usage{
grid_search_eha(cascades, n_jobs, n_edges, params, time_windows)
}
\arguments{
\item{cascades}{an object of class cascade containing node and cascade
information. See \code{\link{as_cascade_long}} and
\code{\link{as_cascade_wide}} for details.}
}
\value{
Returns an object of class \code{data.frame}
}
\description{
This the short description
}
\details{
And some details
}
\examples{

data(cascades)

}
