% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_utilities.R
\name{infer_network}
\alias{infer_network}
\title{Infer diffusion network on time slice of cascades}
\usage{
infer_network(time, time_window, cascades, params, n_edges)
}
\arguments{
\item{cascades}{an object of class cascade containing node and cascade
information. See \code{\link{as_cascade_long}} and
\code{\link{as_cascade_wide}} for details.}
}
\value{
Returns an object of class \code{data.frame}
}
\description{
This the short description
}
\details{
And some details
}
\examples{

data(cascades)

}
