% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eha_utilities.R
\name{make_eha_data}
\alias{make_eha_data}
\title{Create a policy adoption event history data frame}
\usage{
make_eha_data(cascades, networks, decay_parameter, min_time)
}
\arguments{
\item{cascades}{an object of class cascade containing node and cascade
information. See \code{\link{as_cascade_long}} and
\code{\link{as_cascade_wide}} for details.}
}
\value{
Returns an object of class \code{data.frame}
}
\description{
This the short description
}
\details{
And some details
}
\examples{

data(cascades)

}
